配置user redis

+++ internal/dao/relation.go
@@ -2,7 +2,7 @@ package dao
 
 import "simple-demo/internal/model"
 
-func (d *Dao) RelationAction(username string, toUserId uint32, acitonType uint8) error {
+func (d *Dao) RelationAction(username string, toUserId int64, acitonType uint8) error {
 	relation := model.Relation{
 		UserId:     toUserId,
 		UserName:   username,
@@ -11,14 +11,14 @@ func (d *Dao) RelationAction(username string, toUserId uint32, acitonType uint8)
 	return relation.RelationAction(d.engine)
 }
 
-func (d *Dao) FollowList(userId uint32) ([]model.User, error) {
+func (d *Dao) FollowList(userId int64) ([]model.User, error) {
 	relation := model.Relation{
 		FollowerId: userId,
 	}
 	return relation.FollowList(d.engine)
 }
 
-func (d *Dao) FollowerList(userId uint32) ([]model.User, error) {
+func (d *Dao) FollowerList(userId int64) ([]model.User, error) {
 	relation := model.Relation{
 		Id: userId,
 	}
